+++ src/QueryManager.ts
@@ -39,6 +39,12 @@ import {
   GraphQLResult,
   Document,
   FragmentDefinition,
+  // We need to import this here to allow TypeScript to include it in the definition file even
+  // though we don't use it. https://github.com/Microsoft/TypeScript/issues/5711
+  // We need to disable the linter here because TSLint rightfully complains that this is unused.
+  /* tslint:disable */
+  SelectionSet,
+  /* tslint:enable */
 } from 'graphql';
 
 import { print } from 'graphql-tag/printer';
@@ -88,9 +94,9 @@ export type QueryListener = (queryStoreValue: QueryStoreValue) => void;
 export class QueryManager {
   public pollingTimers: {[queryId: string]: NodeJS.Timer | any}; //oddity in Typescript
   public scheduler: QueryScheduler;
+  public store: ApolloStore;
 
   private networkInterface: NetworkInterface;
-  private store: ApolloStore;
   private reduxRootKey: string;
   private queryTransformer: QueryTransformer;
   private queryListeners: { [queryId: string]: QueryListener };
@@ -119,6 +125,11 @@ export class QueryManager {
     observableQuery: ObservableQuery;
     subscriptions: Subscription[];
   } };
+
+  // A map going from the name of a query to an observer issued for it by watchQuery. This is
+  // generally used to refetches for refetchQueries and to update mutation results through
+  // updateQueries.
+  private queryIdsByName: { [queryName: string]: string[] };
 
   constructor({
     networkInterface,
@@ -158,6 +169,7 @@ export class QueryManager {
     this.batcher.start(this.batchInterval);
     this.fetchQueryPromises = {};
     this.observableQueries = {};
+    this.queryIdsByName = {};
 
     // this.store is usually the fake store we get from the Redux middleware API
     // XXX for tests, we sometimes pass in a real Redux store into the QueryManager
@@ -187,6 +199,7 @@ export class QueryManager {
     fragments = [],
     optimisticResponse,
     updateQueries,
+    refetchQueries = [],
   }: {
     mutation: Document,
     variables?: Object,
@@ -194,6 +207,7 @@ export class QueryManager {
     fragments?: FragmentDefinition[],
     optimisticResponse?: Object,
     updateQueries?: MutationQueryReducersMap,
+    refetchQueries?: string[],
   }): Promise<ApolloQueryResult> {
     const mutationId = this.generateQueryId();
 
@@ -256,6 +270,7 @@ export class QueryManager {
             ],
           });
 
+          refetchQueries.forEach((name) => { this.refetchQueryByName(name); });
           resolve(result);
         })
         .catch((err) => {
@@ -313,6 +328,7 @@ export class QueryManager {
                 returnPartialData: options.returnPartialData || options.noFetch,
                 fragmentMap: queryStoreValue.fragmentMap,
               }),
+              loading: queryStoreValue.loading,
             };
 
             if (observer.next) {
@@ -425,9 +441,19 @@ export class QueryManager {
     delete this.fetchQueryPromises[requestId.toString()];
   }
 
-  // Adds an ObservableQuery to this.observableQueries
+  // Adds an ObservableQuery to this.observableQueries and to this.observableQueriesByName.
   public addObservableQuery(queryId: string, observableQuery: ObservableQuery) {
     this.observableQueries[queryId] = { observableQuery, subscriptions: [] };
+
+    // Insert the ObservableQuery into this.observableQueriesByName if the query has a name
+    const queryDef = getQueryDefinition(observableQuery.options.query);
+    if (queryDef.name && queryDef.name.value) {
+      const queryName = getQueryDefinition(observableQuery.options.query).name.value;
+
+      // XXX we may we want to warn the user about query name conflicts in the future
+      this.queryIdsByName[queryName] = this.queryIdsByName[queryName] || [];
+      this.queryIdsByName[queryName].push(observableQuery.queryId);
+    }
   }
 
   // Associates a query subscription with an ObservableQuery in this.observableQueries
@@ -443,7 +469,12 @@ export class QueryManager {
   }
 
   public removeObservableQuery(queryId: string) {
+    const observableQuery = this.observableQueries[queryId].observableQuery;
+    const queryName = getQueryDefinition(observableQuery.options.query).name.value;
     delete this.observableQueries[queryId];
+    this.queryIdsByName[queryName] = this.queryIdsByName[queryName].filter((val) => {
+      return !(observableQuery.queryId === val);
+    });
   }
 
   public resetStore(): void {
@@ -495,6 +526,53 @@ export class QueryManager {
     this.stopQueryInStore(queryId);
   }
 
+  public getQueryWithPreviousResult(queryId: string, isOptimistic = false) {
+    if (!this.observableQueries[queryId]) {
+      throw new Error(`ObservableQuery with this id doesn't exist: ${queryId}`);
+    }
+
+    const observableQuery = this.observableQueries[queryId].observableQuery;
+
+    const queryOptions = observableQuery.options;
+
+    let fragments = queryOptions.fragments;
+    let queryDefinition = getQueryDefinition(queryOptions.query);
+
+    if (this.queryTransformer) {
+      const doc = {
+        kind: 'Document',
+        definitions: [
+          queryDefinition,
+            ...(fragments || []),
+        ],
+      };
+
+      const transformedDoc = applyTransformers(doc, [this.queryTransformer]);
+
+      queryDefinition = getQueryDefinition(transformedDoc);
+      fragments = getFragmentDefinitions(transformedDoc);
+    }
+
+    const previousResult = readSelectionSetFromStore({
+      // In case of an optimistic change, apply reducer on top of the
+      // results including previous optimistic updates. Otherwise, apply it
+      // on top of the real data only.
+      store: isOptimistic ? this.getDataWithOptimisticResults() : this.getApolloState().data,
+      rootId: 'ROOT_QUERY',
+      selectionSet: queryDefinition.selectionSet,
+      variables: queryOptions.variables,
+      returnPartialData: queryOptions.returnPartialData || queryOptions.noFetch,
+      fragmentMap: createFragmentMap(fragments || []),
+    });
+
+    return {
+      previousResult,
+      queryVariables: queryOptions.variables,
+      querySelectionSet: queryDefinition.selectionSet,
+      queryFragments: fragments,
+    };
+  }
+
   private collectResultBehaviorsFromUpdateQueries(
     updateQueries: MutationQueryReducersMap,
     mutationResult: Object,
@@ -505,67 +583,32 @@ export class QueryManager {
     }
     const resultBehaviors = [];
 
-    const observableQueriesByName: { [name: string]: ObservableQuery[] } = {};
-    Object.keys(this.observableQueries).forEach((key) => {
-      const observableQuery = this.observableQueries[key].observableQuery;
-      const queryName = getQueryDefinition(observableQuery.options.query).name.value;
-
-      observableQueriesByName[queryName] =
-        observableQueriesByName[queryName] || [];
-      observableQueriesByName[queryName].push(observableQuery);
-    });
-
     Object.keys(updateQueries).forEach((queryName) => {
       const reducer = updateQueries[queryName];
-      const queries = observableQueriesByName[queryName];
-      if (!queries) {
+      const queryIds = this.queryIdsByName[queryName];
+      if (!queryIds) {
         // XXX should throw an error?
         return;
       }
 
-      queries.forEach((observableQuery) => {
-        const queryOptions = observableQuery.options;
-
-        let fragments = queryOptions.fragments;
-        let queryDefinition = getQueryDefinition(queryOptions.query);
-
-        if (this.queryTransformer) {
-          const doc = {
-            kind: 'Document',
-            definitions: [
-              queryDefinition,
-              ...(fragments || []),
-            ],
-          };
-
-          const transformedDoc = applyTransformers(doc, [this.queryTransformer]);
-
-          queryDefinition = getQueryDefinition(transformedDoc);
-          fragments = getFragmentDefinitions(transformedDoc);
-        }
-
-        const previousResult = readSelectionSetFromStore({
-          // In case of an optimistic change, apply reducer on top of the
-          // results including previous optimistic updates. Otherwise, apply it
-          // on top of the real data only.
-          store: isOptimistic ? this.getDataWithOptimisticResults() : this.getApolloState().data,
-          rootId: 'ROOT_QUERY',
-          selectionSet: queryDefinition.selectionSet,
-          variables: queryOptions.variables,
-          returnPartialData: queryOptions.returnPartialData || queryOptions.noFetch,
-          fragmentMap: createFragmentMap(fragments || []),
-        });
+      queryIds.forEach((queryId) => {
+        const {
+          previousResult,
+          queryVariables,
+          querySelectionSet,
+          queryFragments,
+        } = this.getQueryWithPreviousResult(queryId, isOptimistic);
 
         resultBehaviors.push({
           type: 'QUERY_RESULT',
           newResult: reducer(previousResult, {
             mutationResult,
             queryName,
-            queryVariables: queryOptions.variables,
+            queryVariables,
           }),
-          queryVariables: queryOptions.variables,
-          querySelectionSet: queryDefinition.selectionSet,
-          queryFragments: fragments,
+          queryVariables,
+          querySelectionSet,
+          queryFragments,
         });
       });
     });
@@ -732,7 +775,7 @@ export class QueryManager {
 
             // return a chainable promise
             this.removeFetchQueryPromise(requestId);
-            resolve({ data: resultFromStore });
+            resolve({ data: resultFromStore, loading: false });
           }).catch((error: Error) => {
             this.store.dispatch({
               type: 'APOLLO_QUERY_ERROR',
@@ -751,6 +794,14 @@ export class QueryManager {
     // return a chainable promise
     return new Promise((resolve) => {
       resolve({ data: initialResult });
+    });
+  }
+
+  // Refetches a query given that query's name. Refetches
+  // all ObservableQuery instances associated with the query name.
+  private refetchQueryByName(queryName: string) {
+    this.queryIdsByName[queryName].forEach((queryId) => {
+      this.observableQueries[queryId].observableQuery.refetch();
     });
   }