+++ src/queries/getFromAST.ts
@@ -111,7 +111,7 @@ export interface FragmentMap {
 
 // Utility function that takes a list of fragment definitions and makes a hash out of them
 // that maps the name of the fragment to the fragment definition.
-export function createFragmentMap(fragments: FragmentDefinition[]): FragmentMap {
+export function createFragmentMap(fragments: FragmentDefinition[] = []): FragmentMap {
   const symTable: FragmentMap = {};
   fragments.forEach((fragment) => {
     symTable[fragment.name.value] = fragment;