+++ test/client.ts
@@ -215,7 +215,7 @@ describe('client', () => {
 
     return client.query({ query })
       .then((result) => {
-        assert.deepEqual(result, { data });
+        assert.deepEqual(result.data, data);
       });
   });
 
@@ -259,7 +259,7 @@ describe('client', () => {
 
     return client.query({ query })
       .then((result) => {
-        assert.deepEqual(result, { data });
+        assert.deepEqual(result.data, data);
       });
   });
 
@@ -298,7 +298,7 @@ describe('client', () => {
 
     return client.query({ query })
       .then((result) => {
-        assert.deepEqual(result, { data });
+        assert.deepEqual(result.data, data);
       });
   });
 
@@ -378,7 +378,7 @@ describe('client', () => {
 
     return client.query({ query })
       .then((result) => {
-        assert.deepEqual(result, { data });
+        assert.deepEqual(result.data, data);
         assert.deepEqual(initialState, client.store.getState());
       });
   });
@@ -426,7 +426,7 @@ describe('client', () => {
 
     return client.query({ query })
       .then((result) => {
-        assert.deepEqual(result, { data });
+        assert.deepEqual(result.data, data);
       });
   });
   it('should return errors correctly for a single query', () => {
@@ -862,7 +862,7 @@ describe('client', () => {
 
       return client.query({ query })
         .then((result) => {
-          assert.deepEqual(result, { data });
+          assert.deepEqual(result.data, data);
           assert.deepEqual(client.store.getState()['apollo'].data['1'],
             {
               id: '1',
@@ -893,7 +893,7 @@ describe('client', () => {
 
       return client.query({ query })
         .then((result) => {
-          assert.deepEqual(result, { data });
+          assert.deepEqual(result.data, data);
           assert.deepEqual(store.getState()['apollo'].data['1'],
             {
               id: '1',
@@ -1262,7 +1262,7 @@ describe('client', () => {
         }`);
 
       client.query({ query: queryDoc, fragments: fragmentDefs }).then((result) => {
-        assert.deepEqual(result, { data });
+        assert.deepEqual(result.data, data);
         done();
       });
     });
@@ -1348,7 +1348,7 @@ describe('client', () => {
       const observer = client.watchQuery({ query: queryDoc, fragments: fragmentDefs });
       observer.subscribe({
         next(result) {
-          assert.deepEqual(result, { data });
+          assert.deepEqual(result.data, data);
           done();
         },
       });
@@ -1394,7 +1394,7 @@ describe('client', () => {
         { query: queryDoc, pollInterval: 30, fragments: fragmentDefs});
       const subscription = observer.subscribe({
         next(result) {
-          assert.deepEqual(result, { data });
+          assert.deepEqual(result.data, data);
           subscription.unsubscribe();
           done();
         },