+++ test/tests.ts
@@ -24,5 +24,6 @@ import './batching';
 import './scheduler';
 import './mutationResults';
 import './optimistic';
+import './fetchMore';
 import './scopeQuery';
 import './errors';